feat: add latest features in /ai page and refactor Core Theme Generation logic hooks
