+++ app/ai/components/ai-chat-hero.tsx
@@ -1,12 +1,11 @@
 "use client";
 
 import { HorizontalScrollArea } from "@/components/horizontal-scroll-area";
-import { useAIThemeGeneration } from "@/hooks/use-ai-theme-generation";
+import { useAIGenerateTheme } from "@/hooks/use-ai-generate-theme";
+import { useGuards } from "@/hooks/use-guards";
 import { usePostLoginAction } from "@/hooks/use-post-login-action";
-import { toast } from "@/hooks/use-toast";
-import { authClient } from "@/lib/auth-client";
 import { useAIChatStore } from "@/store/ai-chat-store";
-import { useAuthStore } from "@/store/auth-store";
+import { usePreferencesStore } from "@/store/preferences-store";
 import { AIPromptData } from "@/types/ai";
 import { useRouter } from "next/navigation";
 import { AIChatForm } from "./ai-chat-form";
@@ -14,68 +13,26 @@ import { ChatHeading } from "./chat-heading";
 import { SuggestedPillActions } from "./suggested-pill-actions";
 
 export function AIChatHero() {
-  const { addUserMessage, addAssistantMessage, clearMessages } = useAIChatStore();
+  const { generateTheme } = useAIGenerateTheme();
+  const { checkValidSession, checkValidSubscription } = useGuards();
   const router = useRouter();
-  const { generateTheme } = useAIThemeGeneration();
 
-  const { data: session } = authClient.useSession();
-  const { openAuthDialog } = useAuthStore();
+  const { clearMessages } = useAIChatStore();
+  const { setChatSuggestionsOpen } = usePreferencesStore();
 
   const handleRedirectAndThemeGeneration = async (promptData: AIPromptData | null) => {
-    if (!session) {
-      if (promptData) {
-        openAuthDialog("signup", "AI_GENERATE_FROM_PAGE", { promptData });
-      }
-      return;
-    }
+    if (!checkValidSession("signup", "AI_GENERATE_FROM_PAGE", { promptData })) return;
+    if (!checkValidSubscription()) return;
 
-    // TODO: Check
-    // Ideally we should check the subscription status here, and open the modal if needed...
-
-    if (!promptData) {
-      toast({
-        title: "Error",
-        description: "Failed to generate theme. Please try again.",
-      });
-      return;
-    }
-
-    // Clear the messages when the user starts a chat from the '/ai' page
+    // Clear the messages and open the chat suggestions when the user starts a chat from the '/ai' page
     clearMessages();
-
-    addUserMessage({
-      promptData,
-    });
+    setChatSuggestionsOpen(true);
 
     router.push("/editor/theme?tab=ai");
-
-    const updatedMessages = useAIChatStore.getState().messages;
-    const result = await generateTheme(updatedMessages);
-
-    if (!result) {
-      addAssistantMessage({
-        content: "Failed to generate theme.",
-      });
-      return;
-    }
-
-    addAssistantMessage({
-      content:
-        result?.text ??
-        (result?.theme ? "Here's the theme I generated for you." : "Failed to generate theme."),
-      themeStyles: result?.theme,
-    });
+    generateTheme(promptData);
   };
 
   usePostLoginAction("AI_GENERATE_FROM_PAGE", ({ promptData }) => {
-    if (!promptData) {
-      toast({
-        title: "Error",
-        description: "Failed to generate theme. Please try again.",
-      });
-      return;
-    }
-
     handleRedirectAndThemeGeneration(promptData);
   });
