+++ components/editor/ai/closeable-suggested-pill-actions.tsx
@@ -2,7 +2,7 @@
 
 import { HorizontalScrollArea } from "@/components/horizontal-scroll-area";
 import { Button } from "@/components/ui/button";
-import { useAIThemeGeneration } from "@/hooks/use-ai-theme-generation";
+import { useAIThemeGenerationCore } from "@/hooks/use-ai-theme-generation-core";
 import { usePreferencesStore } from "@/store/preferences-store";
 import { AIPromptData } from "@/types/ai";
 import { createCurrentThemePrompt } from "@/utils/ai/ai-prompt";
@@ -15,7 +15,7 @@ export function ClosableSuggestedPillActions({
 }: {
   onGenerateTheme: (promptData: AIPromptData | null) => Promise<void>;
 }) {
-  const { loading: aiIsGenerating } = useAIThemeGeneration();
+  const { loading: aiIsGenerating } = useAIThemeGenerationCore();
   const { chatSuggestionsOpen, setChatSuggestionsOpen } = usePreferencesStore();
 
   const handleSetPrompt = async (prompt: string) => {
@@ -28,7 +28,7 @@ export function ClosableSuggestedPillActions({
   return (
     <div className="relative flex flex-col items-center justify-center">
       {/* Fade out effect when scrolling */}
-      <div className="via-background/50 from-background pointer-events-none absolute -top-8 right-0 left-0 z-20 h-8 bg-gradient-to-t to-transparent opacity-100 transition-opacity ease-out" />
+      <div className="via-background/50 from-background pointer-events-none absolute -top-6 right-0 left-0 z-20 h-6 bg-gradient-to-t to-transparent opacity-100 transition-opacity ease-out" />
 
       <div className="text-muted-foreground flex w-full items-center justify-between gap-4">
         <h3 className="text-xs">Suggestions</h3>
